CVC4: implement close operations in a way that they can be called several times.

+++ src/org/sosy_lab/java_smt/solvers/cvc4/CVC4AbstractProver.java
@@ -167,8 +167,9 @@ abstract class CVC4AbstractProver<T> implements BasicProverEnvironment<T> {
 
   @Override
   public void close() {
-    Preconditions.checkState(!closed);
-    smtEngine.delete();
-    closed = true;
+    if (!closed) {
+      smtEngine.delete();
+      closed = true;
+    }
   }
 }
